Original change: https://android-review.googlesource.com/c/platform/frameworks/support/+/3766784

+++ busytown/impl/build.sh
@@ -20,7 +20,7 @@ mkdir -p "$DIST_DIR"
 export DIST_DIR="$DIST_DIR"
 if [ "$CHANGE_INFO" != "" ]; then
   cp "$CHANGE_INFO" "$DIST_DIR/"
-  if [ "$MANIFEST" == "" ]; then
+  if [ "$MANIFEST" == "" ] && [ -f "$DIST_DIR/manifest_${BUILD_NUMBER}.xml" ]; then
     mv "$DIST_DIR/manifest_${BUILD_NUMBER}.xml" "$DIST_DIR/manifest.xml"
     export MANIFEST="$DIST_DIR/manifest.xml"
   fi
